add container pattern, bugfix, and add the shopPage to the firebase database
