+++ src/App.js
@@ -68,7 +68,7 @@ class App extends React.Component {
 }
 
 const mapStateToProps = createStructuredSelector({
-  currentUser: selectCurrentUser,
+  currentUser: selectCurrentUser
 })
 
 const mapDispatchToProps = (dispatch) => ({
