+++ src/components/collections-overview/collections-overview.container.jsx
@@ -1,0 +1,21 @@
+import { connect } from 'react-redux'
+import { createStructuredSelector } from 'reselect'
+import { compose } from 'redux'
+import { selectIsCollectionFetching } from '../../redux/shop/shop.selectors'
+import WithSpinner from '../with-spinner/with-spinner.component'
+import CollectionsOverviewComponent from './collections-overview.component'
+import CollectionsOverview from './collections-overview.component'
+
+const mapStateToProps = createStructuredSelector({
+  isLoading: selectIsCollectionFetching,
+})
+
+const CollectionsOverviewContainer = compose(
+    // Will evaluates withSpinner first then passing 
+    // collectionsOverview to withSpinner, then passing
+    // collectionsOverview to connect(mapStateToProps)
+  connect(mapStateToProps),
+  WithSpinner
+)(CollectionsOverview)
+
+export default CollectionsOverviewContainer
