+++ src/components/stripe-button/stripe-button.component.jsx
@@ -1,17 +1,16 @@
 import React from 'react'
 import StripeCheckout from 'react-stripe-checkout'
 
-
 const StripeCheckoutButton = ({ price }) => {
   const priceForStripe = price * 100
   const publishableKey = process.env.REACT_APP_STRIPE_KEY
-  const onToken = token => {
+  const onToken = (token) => {
     console.log(token)
     alert('Payment Successful')
   }
 
   return (
-    <StripeCheckout 
+    <StripeCheckout
       label='Pay Now'
       name='Clothing App'
       billingAddress
