+++ src/pages/collection/collection.container.jsx
@@ -1,0 +1,18 @@
+import { connect } from "react-redux";
+import { compose } from "redux";
+import { createStructuredSelector } from "reselect";
+
+import { selectIsCollectionsLoaded } from "../../redux/shop/shop.selectors";
+import  WithSpinner  from "../../components/with-spinner/with-spinner.component";
+import CollectionPage from './collection.component'
+
+const mapStateToProps = createStructuredSelector({
+    isLoading: state => !selectIsCollectionsLoaded(state)
+});
+
+const CollectionPageContainer = compose(
+    connect(mapStateToProps),
+    WithSpinner
+)(CollectionPage)
+
+export default CollectionPageContainer
