+++ src/redux/shop/shop.actions.js
@@ -1,6 +1,40 @@
-import ShopActionTypes from "./shop.types";
+import ShopActionTypes from './shop.types'
 
-export const updateCollections = collectionsMap => ({
-    type: ShopActionTypes.UPDATE_COLLECTIONS,
-    payload: collectionsMap
+import {
+  firestore,
+  convertCollectionsSnapshotToMap,
+} from '../../firebase/firebase.utils'
+// export const updateCollections = collectionsMap => ({
+//     type: ShopActionTypes.UPDATE_COLLECTIONS,
+//     payload: collectionsMap
+// })
+
+export const fetchCollectionsStart = () => ({
+  type: ShopActionTypes.FETCH_COLLECTIONS_START,
 })
+
+export const fetchCollectionsSucccess = (collectionsMap) => ({
+  type: ShopActionTypes.FETCH_COLLECTIONS_SUCCESS,
+  payload: collectionsMap,
+})
+
+export const fetchCollectionsFailure = (errorMessage) => ({
+  type: ShopActionTypes.FETCH_COLLECTIONS_FAILURE,
+  payload: errorMessage,
+})
+
+export const fetchCollectionsStartAsync = () => {
+  return (dispatch) => {
+    const collectionRef = firestore.collection('collections')
+    dispatch(fetchCollectionsStart())
+
+    collectionRef
+      .get()
+      .then((snapshot) => {
+        const collectionsMap = convertCollectionsSnapshotToMap(snapshot)
+        //   updateCollections(collectionsMap)
+        this.setState({ loading: false })
+      })
+      .catch((error) => dispatch(fetchCollectionsFailure(error.message)))
+  }
+}
