+++ src/redux/shop/shop.selectors.js
@@ -1,4 +1,4 @@
-import { createSelector } from "reselect";
+import { createSelector } from 'reselect'
 import memoize from 'lodash.memoize'
 
 // const COLLECTION_ID_MAP = {
@@ -6,23 +6,35 @@ import memoize from 'lodash.memoize'
 //   sneakers: 2,
 //   jackets: 3,
 //   womens: 4,
-//   mens: 5
+//   mens: 5,
 // }
 
-const selectShop = state => state.shop
+const selectShop = (state) => state.shop
 
 export const selectCollections = createSelector(
   [selectShop],
-  shop => shop.collections
+  (shop) => shop.collections
 )
 
 export const selectCollectionsForPreview = createSelector(
   [selectCollections],
-  collections => collections ?  Object.keys(collections).map(key => collections[key]) : []
+  (collections) =>
+    collections ? Object.keys(collections).map((key) => collections[key]) : []
 )
 
-export const selectCollection = memoize(collectionUrlParam => createSelector(
-  [selectCollections],
-  (collections) => collections ? collections[collectionUrlParam] : null
-  // collections => collections.find(collection => collection.id === COLLECTION_ID_MAP[collectionUrlParam])
-))
+export const selectCollection = memoize((collectionUrlParam) =>
+  createSelector(
+    [selectCollections],
+    (collections) => (collections ? collections[collectionUrlParam] : null)
+  )
+)
+
+export const selectIsCollectionFetching = createSelector(
+  [selectShop],
+  (shop) => shop.isFetching
+)
+
+export const selectIsCollectionsLoaded = createSelector(
+  [selectShop],
+  (shop) => !!shop.collections
+)
