+++ src/redux/shop/shop.types.js
@@ -1,5 +1,7 @@
 const ShopActionTypes = {
-  UPDATE_COLLECTIONS: 'UPDATE_COLLECTIONS',
+  FETCH_COLLECTIONS_START: 'FETCH_COLLECTIONS_START',
+  FETCH_COLLECTIONS_SUCCESS: 'FETCH_COLLECTIONS_SUCCESS',
+  FETCH_COLLECTIONS_FAILURE: 'FETCH_COLLECTIONS_FAILURE'
 }
 
 export default ShopActionTypes
