+++ src/redux/store.js
@@ -1,10 +1,11 @@
 import { createStore, applyMiddleware } from 'redux'
 import { persistStore } from "redux-persist";
 import logger from 'redux-logger'
+import thunk from 'redux-thunk'
 
 import rootReducer from './root-reducer'
 
-const middlewares = []
+const middlewares = [thunk] 
 
 if(process.env.NODE_ENV === 'development') {
   middlewares.push(logger)
